+++ NEWS.md
@@ -1,3 +1,11 @@
+# aplot 0.2.0
+
++ control space between subplots (2023-07-15, Sat, #31)
++ `gglistGrob()` to convert a 'gglist' object to a 'gtable' object (2023-06-26, Mon)
++ `<=` to add ggplot component to each of the plots stored in a "gglist" object (2023-06-26, Mon)
++ mv `theme_no_margin()` to the 'ggfun' package (2023-06-24, Sat)
++ extend `ggsave()` to support `gglist` object (2023-06-05, Mon, #26)
+
 # aplot 0.1.10
 
 + fixed R check for a typo in `grid.draw.aplot()` method (2023-03-08, Wed)